nueva actualizacion, integracion registro y login con roles

+++ src/main/java/com/proyectoTFG/PoyectoTFG/repositories/CiudadRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyectoTFG.PoyectoTFG.entities.Ciudad;
 
 
-public interface CiudadRepository extends JpaRepository<Ciudad, Integer>{
+public interface CiudadRepository extends JpaRepository<Ciudad, Long>{
     
 }
