+++ src/main/java/com/proyectoTFG/PoyectoTFG/repositories/CompraRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyectoTFG.PoyectoTFG.entities.Compra;
 
 
-public interface CompraRepository extends JpaRepository<Compra, Integer>{
+public interface CompraRepository extends JpaRepository<Compra, Long>{
     
 }
