+++ src/main/java/com/proyectoTFG/PoyectoTFG/repositories/PaisRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyectoTFG.PoyectoTFG.entities.Pais;
 
 
-public interface PaisRepository extends JpaRepository<Pais, Integer>{
+public interface PaisRepository extends JpaRepository<Pais, Long>{
     
 }
