+++ src/main/java/com/proyectoTFG/PoyectoTFG/repositories/ProovedorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyectoTFG.PoyectoTFG.entities.Proovedor;
 
 
-public interface ProovedorRepository extends JpaRepository<Proovedor, Integer>{
+public interface ProovedorRepository extends JpaRepository<Proovedor, Long>{
     
 }
