+++ src/main/java/com/proyectoTFG/PoyectoTFG/repositories/ProvinciaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyectoTFG.PoyectoTFG.entities.Provincia;
 
 
-public interface ProvinciaRepository extends JpaRepository<Provincia, Integer>{
+public interface ProvinciaRepository extends JpaRepository<Provincia, Long>{
     
 }
