+++ src/main/java/com/proyectoTFG/PoyectoTFG/services/CategoriaService.java
@@ -7,7 +7,7 @@ import com.proyectoTFG.PoyectoTFG.entities.Categoria;
 public interface CategoriaService {
     
     List<Categoria> findAll();
-    Categoria findById(Integer id);
+    Categoria findById(Long id);
     Categoria save(Categoria categoria);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
