+++ src/main/java/com/proyectoTFG/PoyectoTFG/services/PaisService.java
@@ -7,7 +7,7 @@ import com.proyectoTFG.PoyectoTFG.entities.Pais;
 public interface PaisService {
     
     List<Pais> findAll();
-    Pais findById(Integer id);
+    Pais findById(Long id);
     Pais save(Pais pais);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
