+++ src/main/java/com/proyectoTFG/PoyectoTFG/services/ProvinciaService.java
@@ -7,7 +7,7 @@ import com.proyectoTFG.PoyectoTFG.entities.Provincia;
 public interface ProvinciaService {
     
     List<Provincia> findAll();
-    Provincia findById(Integer id);
+    Provincia findById(Long id);
     Provincia save(Provincia provincia);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
